+++ entrypoints/popup/components/BottomNav.tsx
@@ -1,6 +1,6 @@
 import { ToggleButton, ToggleButtonGroup } from 'react-aria-components';
 import { FaChartSimple, FaCode, FaGear, FaHouseChimney } from 'react-icons/fa6';
-import { useReviewQueueQuery } from '@/hooks/queries/cards';
+import { useReviewQueueQuery } from '@/entrypoints/popup/queries/cards';
 import { useI18n } from '../contexts/I18nContext';
 
 export type ViewId = 'home' | 'card' | 'stats' | 'settings';
