+++ entrypoints/popup/components/__tests__/StreakCounter.test.tsx
@@ -2,7 +2,7 @@
 import { render, screen } from '@testing-library/react';
 import { Rating } from 'ts-fsrs';
 import { describe, expect, it, vi } from 'vitest';
-import { statsQueryKeys } from '@/hooks/queries/stats';
+import { statsQueryKeys } from '@/entrypoints/popup/queries/stats';
 import { sendMessage } from '@/shared/messages';
 import type { DailyStats } from '@/shared/stats';
 import { createDeferred } from '@/test/utils/deferred';
