+++ entrypoints/popup/hooks/__tests__/useTimedConfirmation.test.tsx
