+++ entrypoints/popup/hooks/useTheme.ts
@@ -1,6 +1,6 @@
 import { useEffect, useState } from 'react';
 import type { Theme } from '@/shared/settings';
-import { useSettingsQuery } from './queries/settings';
+import { useSettingsQuery } from '../queries/settings';
 
 export type ResolvedTheme = Exclude<Theme, 'system'>;
