+++ entrypoints/popup/hooks/useTimedConfirmation.ts
