+++ entrypoints/popup/queries/__tests__/notes.test.tsx
