+++ entrypoints/popup/queries/__tests__/settings.test.tsx
