+++ entrypoints/popup/queries/cards.ts
