+++ entrypoints/popup/queries/data.ts
