+++ entrypoints/popup/queries/gist-sync.ts
