+++ entrypoints/popup/queries/notes.ts
