+++ entrypoints/popup/queries/settings.ts
