+++ entrypoints/popup/queries/stats.ts
