+++ entrypoints/popup/views/card/components/CardNotes.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useRef } from 'react';
 import { Button, Label, TextArea, TextField } from 'react-aria-components';
+import { useNoteEditor } from '@/entrypoints/popup/hooks/useNoteEditor';
 import { bounceButton } from '@/entrypoints/popup/styles';
-import { useNoteEditor } from '@/hooks/useNoteEditor';
 import { NOTES_MAX_LENGTH } from '@/shared/notes';
 import { useI18n } from '../../../contexts/I18nContext';
