+++ entrypoints/popup/views/settings/ProblemAutoClearSection.tsx
@@ -1,4 +1,4 @@
-import { useSettingsQuery, useUpdateSettingsMutation } from '@/hooks/queries/settings';
+import { useSettingsQuery, useUpdateSettingsMutation } from '@/entrypoints/popup/queries/settings';
 import { useI18n } from '../../contexts/I18nContext';
 import { SettingsSwitch } from './SettingsSwitch';
